// database/prisma-thinkthank/generated/zod/outputTypeSchemas/ProgramFindManyArgsSchema.ts
export const ProgramSelectSchema: z.ZodType<Prisma.ProgramSelect> = z
    .object({
        id: z.boolean().optional(),
        workspaceId: z.boolean().optional(),
        name: z.boolean().optional(),
        slug: z.boolean().optional(),
        logo: z.boolean().optional(),
        wordmark: z.boolean().optional(),
        brandColor: z.boolean().optional(),
        domain: z.boolean().optional(),
        url: z.boolean().optional(),
        type: z.boolean().optional(),
        cookieLength: z.boolean().optional(),
        commissionAmount: z.boolean().optional(),
        commissionType: z.boolean().optional(),
        commissionDuration: z.boolean().optional(),
        commissionInterval: z.boolean().optional(),
        landerData: z.boolean().optional(),
        createdAt: z.boolean().optional(),
        updatedAt: z.boolean().optional(),
        discounts: z.union([z.boolean(), z.lazy(() => DiscountFindManyArgsSchema)]).optional(),
        workspace: z.union([z.boolean(), z.lazy(() => ProjectArgsSchema)]).optional(),
        primaryDomain: z.union([z.boolean(), z.lazy(() => DomainArgsSchema)]).optional(),
        partners: z.union([z.boolean(), z.lazy(() => ProgramEnrollmentFindManyArgsSchema)]).optional(),
        payouts: z.union([z.boolean(), z.lazy(() => PayoutFindManyArgsSchema)]).optional(),
        sales: z.union([z.boolean(), z.lazy(() => SaleFindManyArgsSchema)]).optional(),
        invoices: z.union([z.boolean(), z.lazy(() => InvoiceFindManyArgsSchema)]).optional(),
        resources: z.union([z.boolean(), z.lazy(() => ProgramResourceFindManyArgsSchema)]).optional(),
        applications: z.union([z.boolean(), z.lazy(() => ProgramApplicationFindManyArgsSchema)]).optional(),
        invites: z.union([z.boolean(), z.lazy(() => ProgramInviteFindManyArgsSchema)]).optional(),
        links: z.union([z.boolean(), z.lazy(() => LinkFindManyArgsSchema)]).optional(),
        _count: z.union([z.boolean(), z.lazy(() => ProgramCountOutputTypeArgsSchema)]).optional()
    })
    .strict()

export const ProgramFindManyArgsSchema: z.ZodType<Prisma.ProgramFindManyArgs> = z
    .object({
        select: ProgramSelectSchema.optional(),
        include: ProgramIncludeSchema.optional(),
        where: ProgramWhereInputSchema.optional(),
        orderBy: z.union([ProgramOrderByWithRelationInputSchema.array(), ProgramOrderByWithRelationInputSchema]).optional(),
        cursor: ProgramWhereUniqueInputSchema.optional(),
        take: z.number().optional(),
        skip: z.number().optional(),
        distinct: z.union([ProgramScalarFieldEnumSchema, ProgramScalarFieldEnumSchema.array()]).optional()
    })
    .strict()